몹 뒤질때 / 총알 이펙트, 모든 버튼 클릭 처리, 스테이지 세부 정보에 일러스트 추가
버그.txt 확인 ㄱㄱ

+++ cTimer.cpp
@@ -1,7 +1,7 @@
 #include "DXUT.h"
 #include "cTimer.h"
 
-cTimer::cTimer(float time, function<void()> func, bool loop, int count)
+cTimer::cTimer(double time, function<void()> func, bool loop, int count)
 	:m_time(time * 1000), func(func), b_loop(loop), m_count(count)
 {
 	m_oldtime = GetTickCount();
